refactor: move most iters
